+++ sdk/storage/azure-storage-file-share/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 12.7.0-beta.1 (Unreleased)
 - Fixed a bug that would cause auth failures when building a client by passing an endpoint which had a sas token with protocol set to https,http
+- Fixed a bug where a custom application id in HttpLogOptions would not be added to the User Agent String.
 
 ## 12.6.0 (2020-08-13)
 - GA release for 2019-12-12 service version